requirements.txt
# Base RAG
chromadb==0.6.3
faiss-cpu==1.10.0
langchain>=0.1.28
langchain-community>=0.0.44
langchain-core>=0.3.47
langchain-ollama>=0.3.0
langchain-experimental
ollama==0.4.7

# Traitement de PDF
marker-pdf==1.6.2
numpy==1.26.4
ocrmypdf==16.10.0
opencv-python==4.11.0.86
pdf2image==1.17.0
pdfminer.six==20250327
PyMuPDF==1.25.5  # fitz
PyPDF2==3.0.1
pytesseract==0.3.13
reportlab==4.3.1

# Embeddings et modèles
onnxruntime==1.21.0
sentence-transformers==4.0.2
torch==2.6.0
transformers==4.50.3

# Utilitaires
fastapi==0.115.12
langdetect==1.0.9
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
unidecode==1.3.8
uvicorn==0.34.0
watchdog==4.0.0

# Graph et détection de communautés
python-igraph>=0.10.6
leidenalg>=0.9.1

# Détection d'objets
ultralytics==8.3.109